Adds Rental dteils styles improvements, adds buil-in pipes and custom pipe
