+++ package.json
@@ -20,6 +20,8 @@
     "@angular/platform-browser-dynamic": "~8.1.0",
     "@angular/router": "~8.1.0",
     "bootstrap": "^4.1.0",
+    "font-awesome": "^4.7.0",
+    "ngx-pipes": "^2.5.6",
     "rxjs": "~6.4.0",
     "tslib": "^1.9.0",
     "zone.js": "~0.9.1"
