+++ src/app/app.module.ts
@@ -7,7 +7,7 @@ import { Routes, RouterModule } from '@angular/router';
 // Check This AppRoutingModule!?!?
 //import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HeaderComponent } from './component/shared/header/header.component';
+import { HeaderComponent } from './common/header/header.component';
 import { RentalComponent } from './rental/rental.component';
 import { TempComponent } from './temp/temp.component';
 import { RentalModule } from './rental/rental.module';
