+++ src/app/common/pipes/uppercase.pipes.ts
@@ -1,0 +1,15 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+
+//Decorator
+@Pipe({
+  name: 'upper'
+})
+
+export class UppercasePipe implements PipeTransform {
+
+  transform(value: string): string {
+    const transformedValue = value.toUpperCase();
+    return transformedValue;
+  }
+}
