+++ src/app/rental/rental-list-item/rental-list-item.component.scss
@@ -1,4 +1,4 @@
-@import '~bootstrap/dist/css/bootstrap.min.css';
+// @import '~bootstrap/dist/css/bootstrap.min.css';
 @import 'variables';
 
 a {
@@ -22,13 +22,13 @@ a {
             font-weight: 500;
             font-size: 12px;
             text-transform: uppercase;
-            color: #16861a;
+
         }
 
         .card-title {
             font-size: 19px;
             margin: 3px 0;
-            color: $main-text-color-gray
+            color: $main-text-color
         }
