+++ src/styles.scss
@@ -1,11 +1,125 @@
 /* You can add global styles to this file, and also import other style files */
 
-
-@import '~bootstrap/dist/css/bootstrap.min.css';
+// @import '~bootstrap/dist/css/bootstrap.min.css';
+// @import '~css-star-rating/scss/star-rating';
 @import 'variables';
+//@import 'rental-detail';
 
 html,
 body,
 p {
-    color: $main-text-color-gray;
+    color: $main-text-color;
+}
+
+.apartement {
+    color: #a72da3;
+}
+
+.condo {
+    color: #d68000;
+}
+
+.house {
+    color: #bf1818;
+}
+
+
+.page-title {
+    font-size: 30px;
+    font-weight: 700;
+    margin-bottom: 20px;
+}
+
+.clickable {
+    cursor: pointer;
+}
+
+.btn-ghost {
+    background-color: transparent;
+    color: $main-color;
+    border: 1px solid $main-color !important;
+    font-size: 12px !important;
+
+    &:hover,
+    &:focus {
+        background-color: transparent !important;
+        color: $main-color !important;
+        border: 1px solid $main-color !important;
+    }
+}
+
+.btn-bwm {
+    background-color: $main-color;
+    border-color: $main-color;
+    color: white;
+    padding: 15px 30px;
+    font-size: 20px;
+
+    &:hover,
+    &:focus,
+    &:active {
+        border-color: $main-color !important;
+        background-color: $main-color !important;
+    }
+}
+
+.bwm-form {
+    color: $main-text-color;
+
+    form {
+        margin-bottom: 10px;
+    }
+
+    .alert {
+        p {
+            margin: 0;
+        }
+    }
+
+    h1 {
+        font-weight: 700;
+        margin-bottom: 30px;
+    }
+
+    .image-container {
+        position: relative;
+
+        img {
+            width: 100%;
+            border-radius: 5px;
+        }
+    }
+
+    label {
+        font-size: 12px;
+    }
+
+    .form-text {
+        font-size: 13px;
+    }
+
+    .catchphrase {
+        text-align: center;
+        font-size: 42px;
+        position: absolute;
+        color: #f1f1f1;
+        padding: 40px;
+        background-color: rgba(0, 0, 0, 0.5);
+        margin-top: 50px;
+        letter-spacing: 1px;
+    }
+}
+
+.inline {
+    display: inline-block;
+}
+
+.editable-component {
+    .btn {
+        &.btn-editable {
+            margin-left: 5px;
+            color: white;
+            font-size: 12px;
+        }
+    }
 }
