+++ src/styles/variables.scss
@@ -1,3 +1,3 @@
 $main-color: #F44336;
-$main-text-color-gray: #484848;
-$light-gray: #e9e9e9;
+$main-text-color: #484848;
+$light-gray-color: #e9e9e9;
